Move Settings interfaces to same file
