--- static/settings.interfaces.ts
@@ -1,72 +1,0 @@
-// Copyright (c) 2021, Compiler Explorer Authors
-// All rights reserved.
-//
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions are met:
-//
-//     * Redistributions of source code must retain the above copyright notice,
-//       this list of conditions and the following disclaimer.
-//     * Redistributions in binary form must reproduce the above copyright
-//       notice, this list of conditions and the following disclaimer in the
-//       documentation and/or other materials provided with the distribution.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-// ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-// LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-// CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-// SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-// INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-// CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-// ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-// POSSIBILITY OF SUCH DAMAGE.
-
-type ColourScheme =
-    | 'rainbow'
-    | 'rainbow2'
-    | 'earth'
-    | 'green-blue'
-    | 'gray-shade'
-    | 'rainbow-dark';
-
-export type FormatBase =
-    | 'Google'
-    | 'LLVM'
-    | 'Mozilla'
-    | 'Chromium'
-    | 'WebKit'
-    | 'Microsoft'
-    | 'GNU';
-
-export interface SiteSettings {
-    autoCloseBrackets: boolean;
-    autoIndent: boolean;
-    allowStoreCodeDebug: boolean;
-    alwaysEnableAllSchemes: boolean;
-    colouriseAsm: boolean;
-    colourScheme: ColourScheme;
-    compileOnChange: boolean;
-    // TODO(supergrecko): make this more precise
-    defaultLanguage: string;
-    delayAfterChange: number;
-    enableCodeLens: boolean;
-    enableCommunityAds: boolean
-    enableCtrlS: boolean;
-    enableCtrlStree: boolean;
-    editorsFFont: string
-    editorsFLigatures: boolean;
-    formatBase: FormatBase;
-    hoverShowAsmDoc: boolean;
-    hoverShowSource: boolean;
-    keepSourcesOnLangChange: boolean;
-    newEditorLastLang: boolean;
-    showMinimap: boolean;
-    showQuickSuggestions: boolean;
-    tabWidth: number;
-    theme: 'default' | 'dark';
-    useCustomContextMenu: boolean;
-    useSpaces: boolean;
-    useVim: boolean;
-    wordWrap: boolean;
-}
